+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added: `at-rule-blockless-newline-after` rule.
 - Fixed: `font-weight-notation` now ignore `inherit` value.
 - Fixed: `selector-list-comma-*` rules no longer deliver false positives inside functional notation.
+- Fixed: `indentation` treats `@nest` at-rules like regular rules with selectors.
 
 # 5.1.0