+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # HEAD
 
 - Added: `no-indistinguishable-colors` rule.
-- Added: `at-rule-blockless-newline-after` rule.
+- Added: `at-rule-semicolon-newline-after` rule.
 - Added: `stylelint-disable-reason` rule.
 - Fixed: `no-duplicate-selectors` contextualizes selectors by all at-rules, not just media queries.
 - Fixed: `declaration-bang-space-*` understands arbitrary bang declarations (e.g. `!default`).