+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # HEAD
 
-- Added: `stylelint-selector-no-qualifying-type` rule.
+- Added: `selector-no-qualifying-type` rule.
 - Added: `no-indistinguishable-colors` rule.
 - Added: `at-rule-semicolon-newline-after` rule.
 - Added: `stylelint-disable-reason` rule.